pyproject.toml
[project]
name = "spellcrafting"
version = "0.1.0"
description = "LLMs as a Python language feature"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Sam McVicker"}
]
keywords = ["llm", "ai", "pydantic", "structured-output", "anthropic", "openai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-ai>=1.38.0,<2.0.0",
    "tomli>=2.0.0;python_version<'3.11'",
]

[project.urls]
Homepage = "https://github.com/sammcvicker/spellcrafting"
Documentation = "https://github.com/sammcvicker/spellcrafting"
Repository = "https://github.com/sammcvicker/spellcrafting"

[project.optional-dependencies]
otel = [
    "opentelemetry-api>=1.0.0",
    "opentelemetry-sdk>=1.0.0",
]
logfire = [
    "logfire>=0.1.0",
]
datadog = [
    "ddtrace>=2.0.0",
]
dotenv = [
    "python-dotenv>=1.0.0",
]
all = [
    "opentelemetry-api>=1.0.0",
    "opentelemetry-sdk>=1.0.0",
    "logfire>=0.1.0",
    "ddtrace>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/spellcrafting"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "smoke: tests that hit real LLM APIs",
    "internal: tests that check internal implementation details",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
]
